move assertUser to user utils
