+++ src/index.js
@@ -13,11 +13,10 @@ import { formatError } from 'graphql';
 import checkHeaders from './checkHeaders';
 import schema from './graphql/schema';
 import DataLoaders from './graphql/dataLoaders';
-import { AUTH_ERROR_MSG } from './graphql/util';
 import CookieStore from './CookieStore';
 import { loginRouter, authRouter } from './auth';
 import rollbar from './rollbarInstance';
-import { createOrUpdateUser } from './util/user';
+import { AUTH_ERROR_MSG, createOrUpdateUser } from './util/user';
 
 const app = new Koa();
 const router = Router();
